lib: rename files for fdn8_000
