+++ src/clover_sketches_lib/lib/fdn/draw_fdn8_000.hpp
@@ -4,6 +4,6 @@
 // Copyright (C) 2025  Rob W. Albus
 // Licensed under the GPLv3. See LICENSE for details.
 
-#include "lib/fdn/fdn.hpp"
+#include "lib/fdn/fdn8_000.hpp"
 
 void draw_fdn8_000(const char* id, fdn8_000* fdn_L, fdn8_000* fdn_R);
