+++ src/sketches/019_composition/views/views.cpp
@@ -4,7 +4,7 @@
 
 #include "imgui.h"
 
-#include "lib/fdn/draw_fdn.hpp"
+#include "lib/fdn/draw_fdn8_000.hpp"
 #include "lib/kick_drum/draw_kick_drum.hpp"
 #include "lib/peq/draw_peq.hpp"
