+++ src/sketches/022_sd_reso_chords/composition/graph.hpp
@@ -12,7 +12,7 @@
 
 #include "lib/echo/echo.hpp"
 #include "lib/env_bp/env_bp.hpp"
-#include "lib/fdn/fdn.hpp"
+#include "lib/fdn/fdn8_000.hpp"
 #include "lib/kick_drum/kick_drum.hpp"
 #include "lib/mixer/mixer.hpp"
 #include "lib/peq/peq.hpp"
